Sample console app now runs the DB setup
* The console app lhm.test is renames to lhm.net.sample (this will let
  us have an lhm.net.test with actual tests
* Change the name of the sample sql script and removed 'go' so it can be
 executed
* Run sampledatabase file at start of console app
