+++ src/lhm.net.sample/Properties/AssemblyInfo.cs
